Transferred the method to the Done on editor method
