+++ app/src/main/java/com/example/devhub/Utils/DoneOnEditorActionListener.java
@@ -1,4 +1,20 @@
 package com.example.devhub.Utils;
 
-public class DoneOnEditorActionListener {
+import android.content.Context;
+import android.view.KeyEvent;
+import android.view.inputmethod.EditorInfo;
+import android.view.inputmethod.InputMethodManager;
+import android.widget.TextView;
+
+public class DoneOnEditorActionListener implements TextView.OnEditorActionListener {
+    @Override
+    public boolean onEditorAction(TextView textView, int actionId, KeyEvent keyEvent) {
+
+        if (actionId == EditorInfo.IME_ACTION_DONE) {
+            InputMethodManager imm = (InputMethodManager)textView.getContext().getSystemService(Context.INPUT_METHOD_SERVICE);
+            imm.hideSoftInputFromWindow(textView.getWindowToken(), 0);
+            return true;
+        }
+        return false;
+    }
 }
